+++ activerecord/lib/rails/generators/active_record/application_record/templates/application_record.rb
