+++ activerecord/lib/rails/generators/active_record/model/model_generator.rb
@@ -23,13 +23,11 @@ module ActiveRecord
       end
 
       def create_model_file
-        generate_application_record
         template "model.rb", File.join("app/models", class_path, "#{file_name}.rb")
       end
 
       def create_module_file
         return if regular_class_path.empty?
-        generate_application_record
         template "module.rb", File.join("app/models", "#{class_path.join('/')}.rb") if behavior == :invoke
       end
 
@@ -41,30 +39,9 @@ module ActiveRecord
           attributes.select { |a| !a.reference? && a.has_index? }
         end
 
-        # FIXME: Change this file to a symlink once RubyGems 2.5.0 is required.
-        def generate_application_record
-          if behavior == :invoke && !application_record_exist?
-            template "application_record.rb", application_record_file_name
-          end
-        end
-
         # Used by the migration template to determine the parent name of the model
         def parent_class_name
           options[:parent] || "ApplicationRecord"
-        end
-
-        def application_record_exist?
-          file_exist = nil
-          in_root { file_exist = File.exist?(application_record_file_name) }
-          file_exist
-        end
-
-        def application_record_file_name
-          @application_record_file_name ||= if mountable_engine?
-            "app/models/#{namespaced_path}/application_record.rb"
-          else
-            "app/models/application_record.rb"
-          end
         end
     end
   end
