+++ railties/test/generators/model_generator_test.rb
@@ -6,14 +6,6 @@ class ModelGeneratorTest < Rails::Generators::TestCase
   include GeneratorsTestHelper
   arguments %w(Account name:string age:integer)
 
-  def test_application_record_skeleton_is_created
-    run_generator
-    assert_file "app/models/application_record.rb" do |record|
-      assert_match(/class ApplicationRecord < ActiveRecord::Base/, record)
-      assert_match(/self\.abstract_class = true/, record)
-    end
-  end
-
   def test_help_shows_invoked_generators_options
     content = run_generator ["--help"]
     assert_match(/ActiveRecord options:/, content)
@@ -41,17 +33,6 @@ class ModelGeneratorTest < Rails::Generators::TestCase
     run_generator ["account", "--parent", "Admin::Account"]
     assert_file "app/models/account.rb", /class Account < Admin::Account/
     assert_no_migration "db/migrate/create_accounts.rb"
-  end
-
-  def test_model_with_existent_application_record
-    mkdir_p "#{destination_root}/app/models"
-    touch "#{destination_root}/app/models/application_record.rb"
-
-    Dir.chdir(destination_root) do
-      run_generator ["account"]
-    end
-
-    assert_file "app/models/account.rb", /class Account < ApplicationRecord/
   end
 
   def test_plural_names_are_singularized
