+++ railties/test/generators/plugin_generator_test.rb
@@ -679,20 +679,6 @@ class PluginGeneratorTest < Rails::Generators::TestCase
     assert_file "app/models/bukkits/article.rb", /class Article < ApplicationRecord/
   end
 
-  def test_generate_application_record_when_does_not_exist_in_mountable_engine
-    run_generator [destination_root, "--mountable"]
-    FileUtils.rm "#{destination_root}/app/models/bukkits/application_record.rb"
-    capture(:stdout) do
-      `#{destination_root}/bin/rails g model article`
-    end
-
-    assert_file "#{destination_root}/app/models/bukkits/application_record.rb" do |record|
-      assert_match(/module Bukkits/, record)
-      assert_match(/class ApplicationRecord < ActiveRecord::Base/, record)
-      assert_match(/self\.abstract_class = true/, record)
-    end
-  end
-
   def test_generate_application_mailer_when_does_not_exist_in_mountable_engine
     run_generator [destination_root, "--mountable"]
     FileUtils.rm "#{destination_root}/app/mailers/bukkits/application_mailer.rb"
