+++ railties/test/generators_test.rb
@@ -124,7 +124,7 @@ class GeneratorsTest < Rails::Generators::TestCase
 
   def test_rails_generators_help_does_not_include_app_nor_plugin_new
     output = capture(:stdout) { Rails::Generators.help }
-    assert_no_match(/app/, output)
+    assert_no_match(/app\W/, output)
     assert_no_match(/[^:]plugin/, output)
   end
